requirements-dev.txt
2to3==1.0
alembic==1.4.2
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
emoji==0.5.4
flake8==3.8.3
Flask==1.1.2
grpcio==1.30.0
grpcio-tools==1.30.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
py-cpuinfo==7.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==5.4.3
pytest-benchmark==3.2.3
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
regex==2020.6.8
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
